Added Controller, and user input to populate adventurer details

+++ MazeGameDomain/Services/DecisionTrees/IceCavern.cs
@@ -22,7 +22,7 @@ namespace MazeGameDomain.Services.DecisionTrees
 
                 ProcessPhase = async _ =>
                 {
-                    if (adventurerDetail.Class == Class.Magician &&
+                    if (adventurerDetail.Class == (int)Class.Magician &&
                         adventurerDetail.Specialisation == Specialisation.IceMage)
                     {
                         return true;
@@ -41,7 +41,7 @@ namespace MazeGameDomain.Services.DecisionTrees
 
                 ProcessPhase = async _ =>
                 {
-                    if (adventurerDetail.Class == Class.Magician &&
+                    if (adventurerDetail.Class == (int)Class.Magician &&
                         adventurerDetail.Specialisation == Specialisation.IceMage)
                     {
                         return true;
